requirements_evaluation.txt
# 评估脚本依赖
torch>=1.9.0
torchvision>=0.10.0
torchmetrics>=0.7.0
numpy>=1.21.0
Pillow>=8.3.0
pytorch-fid>=0.3.0
pandas>=1.3.0
pyyaml>=5.4.0

# 可选依赖（用于更好的性能）
tqdm>=4.62.0
matplotlib>=3.4.0  # 用于可视化结果
